REFACTOR: move github to own module
